Decouple destruction method
