+++ qbit-core/src/commonMain/kotlin/qbit/schema/SchemaDsl.kt
@@ -1,11 +1,15 @@
 package qbit.schema
 
 import qbit.api.model.Attr
+import qbit.api.model.QRef
 import qbit.factorization.attrName
-import qbit.factorization.schemaFor
+import qbit.factorization.collectionTypes
+import qbit.factorization.types
+import qbit.factorization.valueTypes
 import qbit.reflection.default
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty0
+import kotlin.reflect.KProperty1
 
 fun schema(body: SchemaBuilder.() -> Unit): List<Attr<Any>> {
     val scb = SchemaBuilder()
@@ -37,4 +41,25 @@ class EntityBuilder(private val type: KClass<*>) {
         uniqueProps.add(type.attrName(prop))
     }
 
+}
+
+private fun schemaFor(type: KClass<*>, unique: Set<String>): Collection<Attr<Any>> {
+    val getters = type.members.filterIsInstance<KProperty1<*, *>>()
+    val (ids, attrs) = getters.partition { it.name == "id" && it.returnType.classifier == Long::class }
+    val id = ids.firstOrNull()
+    id ?: throw IllegalArgumentException("Type $type does not contains id: Long property")
+    return attrs.map {
+        when (it.returnType.classifier) {
+            in valueTypes -> Attr(null, type.attrName(it), types.getValue(it.returnType.classifier as KClass<*>).code, type.attrName(it) in unique, false)
+            in collectionTypes -> {
+                when (val valueType = it.returnType.arguments[0].type!!.classifier as KClass<*>) {
+                    in valueTypes -> Attr(null, type.attrName(it), types.getValue(valueType).list().code, type.attrName(it) in unique, true)
+                    else -> Attr<Any>(null, type.attrName(it), QRef.list().code, type.attrName(it) in unique, true)
+                }
+            }
+            else -> {
+                Attr(null, type.attrName(it), QRef.code, type.attrName(it) in unique, false)
+            }
+        }
+    }
 }
